+++ flexx/ui/layouts/__init__.py
@@ -1,6 +1,8 @@
 """ Namespace for all layout widgets.
 """
 
+from .._widget import Widget
+
 from ._layout import Layout
 from ._box import BoxLayout, HBox, VBox, BoxPanel
 from ._split import SplitPanel
