+++ flexx/ui/layouts/_box.py
@@ -73,7 +73,7 @@ Interactive example:
 
 """
 
-from .. import react
+from ... import react
 from . import Widget, Layout
