+++ flexx/ui/layouts/_dock.py
@@ -2,7 +2,7 @@
 The dockpanel layoujt widget.
 """
 
-from .. import react
+from ... import react
 from . import Widget, Layout
