+++ flexx/ui/layouts/_grid.py
@@ -2,7 +2,7 @@
 
 """
 
-from .. import react
+from ... import react
 from . import Widget, Layout
