+++ flexx/ui/layouts/_layout.py
@@ -1,8 +1,6 @@
 """ Layout widgets
 """
 
-from .. import react
-
 from . import Widget
