+++ flexx/ui/layouts/_pinboard.py
@@ -14,7 +14,7 @@ Example:
 
 """
 
-from .. import react
+from ... import react
 from . import Widget, Layout
