+++ flexx/ui/layouts/_stack.py
@@ -2,7 +2,7 @@
 
 """
 
-from .. import react
+from ... import react
 from . import Widget, Layout
