+++ flexx/ui/layouts/_tabs.py
@@ -2,7 +2,7 @@
 The tabbar widget.
 """
 
-from .. import react
+from ... import react
 from . import Widget, Layout
