+++ flexx/ui/widgets/__init__.py
@@ -8,7 +8,7 @@ from ._button import Button
 from ._slider import Slider
 from ._lineedit import LineEdit
 from ._label import Label
-from ._panel import GroupWidget
+from ._group import GroupWidget
 from ._progressbar import ProgressBar
 from ._plotwidget import PlotWidget
 from ._iframe import IFrame
