+++ flexx/ui/widgets/_button.py
@@ -29,7 +29,7 @@ Example with interaction:
 
 """
 
-from .. import react
+from ... import react
 from . import Widget
