+++ flexx/ui/widgets/_group.py
@@ -30,7 +30,7 @@ Interactive example:
 
 """
 
-from .. import react
+from ... import react
 from . import Widget
