+++ flexx/ui/widgets/_iframe.py
@@ -1,4 +1,4 @@
-from .. import react
+from ... import react
 from . import Widget
