+++ flexx/ui/widgets/_plotwidget.py
@@ -43,7 +43,7 @@ Interactive example:
                 self.plot.ydata(ydata)
 """
 
-from .. import react
+from ... import react
 from . import Widget
